Various performance improvements, mostly focused on vectorizing stuff within numpy/pandas.

Initial runtime:

```
timebudget report per findEquilibrium cycle...
                   fPrime: 113.2%  1264.92ms/cyc @     0.0 calls/cyc
          findEquilibrium: 100.0%  1117.40ms/cyc @     1.0 calls/cyc
                        g:  94.5%  1056.02ms/cyc @    12.7 calls/cyc
               supplySide:  46.3%   517.91ms/cyc @    12.7 calls/cyc
               demandSide:  42.4%   473.62ms/cyc @    12.7 calls/cyc
                 __init__:  11.4%   127.26ms/cyc @     0.0 calls/cyc
                readFiles:   7.0%    78.17ms/cyc @     0.0 calls/cyc
  importOriginDestination:   5.5%    61.84ms/cyc @     0.0 calls/cyc
 initializeAllTimePeriods:   3.8%    42.78ms/cyc @     0.0 calls/cyc
         importPopulation:   1.4%    15.14ms/cyc @     0.0 calls/cyc
   calculateAccessibility:   0.2%     2.06ms/cyc @     0.0 calls/cyc
 importTransitionMatrices:   0.1%     1.18ms/cyc @     0.0 calls/cyc
[ 9.71198e+07 -2.53462e+05 -8.38087e+06 -6.82591e+08 -1.00687e+07]
timebudget report...
                   fPrime: 182148.47ms for      1 calls
          findEquilibrium: 1117.40ms for    144 calls
                        g:   83.42ms for   1823 calls
               supplySide:   40.91ms for   1823 calls
               demandSide:   37.41ms for   1823 calls
                 __init__: 18325.16ms for      1 calls
                readFiles: 11256.48ms for      1 calls
  importOriginDestination: 8904.98ms for      1 calls
 initializeAllTimePeriods: 6159.66ms for      1 calls
         importPopulation: 2180.56ms for      1 calls
   calculateAccessibility:   49.34ms for      6 calls
 importTransitionMatrices:  170.37ms for      1 calls
```
Current (improved) runtime
```timebudget report per findEquilibrium cycle...
                   fPrime: 113.5%  1133.00ms/cyc @     0.0 calls/cyc
          findEquilibrium: 100.0%   998.51ms/cyc @     1.0 calls/cyc
                        g:  94.6%   944.22ms/cyc @    12.3 calls/cyc
               supplySide:  46.0%   459.36ms/cyc @    12.3 calls/cyc
               demandSide:  42.9%   428.20ms/cyc @    12.3 calls/cyc
                 __init__:   3.8%    37.65ms/cyc @     0.0 calls/cyc
 initializeAllTimePeriods:   1.7%    16.58ms/cyc @     0.0 calls/cyc
                readFiles:   1.6%    15.58ms/cyc @     0.0 calls/cyc
         importPopulation:   1.4%    13.81ms/cyc @     0.0 calls/cyc
   calculateAccessibility:   0.2%     1.87ms/cyc @     0.0 calls/cyc
 importTransitionMatrices:   0.1%     1.08ms/cyc @     0.0 calls/cyc
  importOriginDestination:   0.1%     0.53ms/cyc @     0.0 calls/cyc
[ 9.44211e+07 -2.58733e+05 -8.21041e+06 -6.59131e+08 -9.81235e+06]
timebudget report...
                   fPrime: 163151.41ms for      1 calls
          findEquilibrium:  998.51ms for    144 calls
                        g:   76.69ms for   1773 calls
               supplySide:   37.31ms for   1773 calls
               demandSide:   34.78ms for   1773 calls
                 __init__: 5422.13ms for      1 calls
 initializeAllTimePeriods: 2387.11ms for      1 calls
                readFiles: 2243.69ms for      1 calls
         importPopulation: 1988.11ms for      1 calls
   calculateAccessibility:   44.91ms for      6 calls
 importTransitionMatrices:  155.16ms for      1 calls
  importOriginDestination:   76.02ms for      1 calls

```